01_HDFS/csv_to_parquet/requirements.txt
PyYAML
hdfs==2.5.8
pandas==1.0.3
pyarrow==0.11.0